updated audio.c for some minor mistakes

+++ audio.c
@@ -80,14 +80,14 @@ const short** fourty_five_deg_hrtf_right = (const short*)hrtf_right_matrix;
 int* convolve(const short* audio_word_array, int* result) {
   if (left_index_counter < HRTF_LENGTH) {
     for (int i = 0; i <= left_index_counter; i++) {
-      result[0] += (audio_word_array[left_index_counter - i] * hrtf_left_matrix[10][i]);
-      result[1] += (audio_word_array[right_index_counter - i] * hrtf_right_matrix[10][i]);
+      result[0] += (int) (audio_word_array[left_index_counter - i] * hrtf_left_matrix[10][i]);
+      result[1] += (int) (audio_word_array[right_index_counter - i] * hrtf_right_matrix[10][i]);
     }
   } else {
     for (int i = 0; i <= (HRTF_LENGTH - 1); i++) {
-      result[0] += (audio_word_array[left_index_counter - i] *
+      result[0] += (int) (audio_word_array[left_index_counter - i] *
                     hrtf_left_matrix[10][i]);
-      result[1] += (audio_word_array[left_index_counter - i] *
+      result[1] += (int) (audio_word_array[left_index_counter - i] *
                     hrtf_right_matrix[10][i]);
     }
   }
